requirements.txt
torch
torchvision
numpy
huggingface_hub
pillow
tqdm
opencv-python
timm
